feat: refactoring, separated pathnames and fetchdata (#165)
refactoring, separated pathnames and fetchdata
